cnn_pipeline.py saving to folders according to configfile added

+++ loss_plotter.py
@@ -1,7 +1,7 @@
 import os
 import json
 from matplotlib import pyplot as plt
-
+from pathlib import Path
 
 if os.name == "nt":
     from config import WINDOWS_PATHS as PATHS
@@ -17,12 +17,13 @@ for item in PATHS["PATH_RESULTS"].iterdir():
 for directory in result_dirs:
     json_files = list(directory.glob("*.json"))
     print(f"Processing {directory}")
-    directory.joinpath("imgs").mkdir()
+    directory.joinpath("imgs").mkdir(exist_ok=True)
     for json_file in json_files:
-        with open(json_file, "r") as f:
-            data = json.load(f)
-            plt.figure()
-            plt.plot(data["loss"])
-            plt.plot(data["val_loss"], "r")
-            plt.savefig(directory.joinpath("imgs", json_file.stem + ".png"))
-            plt.close()
+        if not directory.joinpath("imgs", json_file.stem + ".png").exists():
+            with open(json_file, "r") as f:
+                data = json.load(f)
+                plt.figure()
+                plt.plot(data["loss"])
+                plt.plot(data["val_loss"], "r")
+                plt.savefig(directory.joinpath("imgs", json_file.stem + ".png"))
+                plt.close()
